Updates to working versions
There is still a bug/design challenge when two MATLAB models are polling
and doing GETs for each other's variables.  A race condition ensues
since MATLAB is single threaded.  We need to figure out a way to get
around this, either by caching or by redesigning. Or by reimplementing
MATLAB to be multithreaded.